requirements.txt
android_backup==0.2.0
appdirs==1.4.4
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
construct==2.10.70
croniter==2.0.5
cryptography==42.0.5
defusedxml==0.7.1
idna==3.7
ifaddr==0.2.0
micloud==0.6
Naked==0.1.32
netifaces==0.11.0
pillow==10.3.0
protobuf==5.26.1
pycparser==2.22
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-miio==0.5.12
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
six==1.16.0
tqdm==4.66.2
tzlocal==5.2
urllib3==2.2.1
vacuum-map-parser-base==0.1.3
zeroconf==0.132.2
